feat(extension, app): change password screen with a temporary location
