+++ apps/extension/src/App/Settings/ChangePassword/index.ts
@@ -1,0 +1,1 @@
+export * from "./ChangePassword";
