--- apps/extension/src/App/Settings/ExtraSettings/ResetPassword/index.ts
@@ -1,1 +1,0 @@
-export { default as ResetPassword } from "./ResetPassword";
